requirements.txt
openai==2.24.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
reportlab==4.4.10
requests==2.32.5
streamlit==1.50.0